requirements.txt
discord.py
requests
python-dotenv
requests
flask
humanize